Create README.md
About my project University/college/ school egistration Form is an HTML and CSS-based application with a graphical interface (GUI).
Its aim is to collect various data from new students, new teachers, new teaching staff, new non-teaching staff, and new School-Management.
The data collected on paper or in physical form can be easily lost or misplaced. But as they are filling in the data in digital 
form, it would not be lost easily or misplaced.